% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{as_tibble.summrt_summary}
\alias{as_tibble.summrt_summary}
\title{Coerce summrt_summary object to a tibble}
\usage{
\method{as_tibble}{summrt_summary}(x, add_pkg_id = FALSE, ...)
}
\arguments{
\item{x}{A \code{summrt_summary} object}

\item{add_pkg_id}{Logical. Should the \code{pkg} field be included as a column.}

\item{...}{ignored.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Coerce summrt_summary object to a tibble
}
\examples{
ex <- readRDS(system.file(
  "extdata", "EpiEstim_example.rds", package = "summrt"
))
summ_ex <- summarize_rtestimate(ex)
as_tibble(summ_ex)
as_tibble(summ_ex, add_pkg_id = TRUE)
}
