% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{autoplot.summrt_summary}
\alias{autoplot.summrt_summary}
\title{Plot a summrt_summary object}
\usage{
\method{autoplot}{summrt_summary}(x, color = "dodgerblue4", add_reference_line = TRUE, ...)
}
\arguments{
\item{x}{A \code{summrt_summary} object}

\item{color}{A character string giving a color}

\item{add_reference_line}{Logical. Display a horizontal line at 1.}

\item{...}{not used.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Plot a summrt_summary object
}
\examples{
ex <- readRDS(system.file(
  "extdata", "EpiEstim_example.rds", package = "summrt"
))
summ_ex <- summarize_rtestimate(ex)
autoplot(summ_ex)
autoplot(summ_ex) + coord_cartesian(ylim = c(0, 2))
}
