% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R, R/tibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

