% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rtestimate.R
\name{summarize_rtestimate}
\alias{summarize_rtestimate}
\alias{summarize_rtestimate.default}
\alias{summarize_rtestimate.cv_poisson_rt}
\alias{summarize_rtestimate.poisson_rt}
\alias{summarize_rtestimate.epinow}
\alias{summarize_rtestimate.estimate_R}
\alias{summarize_rtestimate.Rt}
\title{Extract Rt estimation from a model fit}
\usage{
summarize_rtestimate(x, level = 0.95, ..., notes = "")

\method{summarize_rtestimate}{default}(x, level = 0.95, ..., notes = "")

\method{summarize_rtestimate}{cv_poisson_rt}(
  x,
  level = 0.95,
  lambda = c("lambda.1se", "lambda.min"),
  ...,
  notes = "cv_poisson_rt"
)

\method{summarize_rtestimate}{poisson_rt}(x, level = 0.95, lambda = NULL, ..., notes = "poisson_rt")

\method{summarize_rtestimate}{epinow}(x, level = 0.95, ..., notes = "")

\method{summarize_rtestimate}{estimate_R}(x, level = 0.95, ..., notes = "")

\method{summarize_rtestimate}{Rt}(x, level = 0.95, ..., notes = "")
}
\arguments{
\item{x}{Object to extract Rt from.}

\item{level}{Confidence level for the confidence interval.}

\item{...}{Additional arguments passed to methods.}

\item{notes}{String. Optional notes to add to the summary.}

\item{lambda}{The Poisson parameter (\code{cv_poisson_rt}).}
}
\description{
Extract Rt estimation from a model fit
}
\details{
The \code{estimate_R} method is for the \code{EpiEstim} package. Currently,
only levels in 50\%, 90\% and 95\% confidence levels are allowed.

The \code{Rt} method is for the \code{EpiLPS} package.
}
