% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{as_tibble.summrt_summary}
\alias{as_tibble.summrt_summary}
\title{Coerce summrt_summary object to a tibble}
\usage{
\method{as_tibble}{summrt_summary}(x, add_pkg_id = FALSE, ...)
}
\arguments{
\item{x}{A \code{summrt_summary} object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
A set of name-value pairs. These arguments are
processed with \code{\link[rlang:defusing-advanced]{rlang::quos()}} and support unquote via \code{\link{!!}} and
unquote-splice via \code{\link{!!!}}. Use \verb{:=} to create columns that start with a dot.

Arguments are evaluated sequentially.
You can refer to previously created elements directly or using the \link{.data}
pronoun.
To refer explicitly to objects in the calling environment, use \code{\link{!!}} or
\link{.env}, e.g. \code{!!.data} or \code{.env$.data} for the special case of an object
named \code{.data}.}

\item{.add_pkg_id}{Logical. Should the \code{pkg} field be included as a column.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Coerce summrt_summary object to a tibble
}
\examples{
ex <- readRDS(system.file(
  "extdata", "EpiEstim_example.rds", package = "summrt"
))
summ_ex <- summarize_rtestimate(ex)
as_tibble(summ_ex)
as_tibble(summ_ex, .add_pkg_id = TRUE)
}
